shut up compiler warning

+++ src/YutilsCpp/shape.cpp
@@ -25,7 +25,7 @@ Shape::Shape()
 std::tuple<double, double, double, double> Shape::bounding(std::string &shape)
 {
     // Bounding data
-    double x0(0.f), y0(0.f), x1(0.f), y1(0.f);
+    double x0(0.), y0(0.), x1(0.), y1(0.);
     uint8_t flag(0);
     std::function<std::pair<double, double>(double, double, std::string&)> flt(
         [&](double x, double y, std::string &pointType)
@@ -74,10 +74,10 @@ std::string Shape::filter(std::string &shape,
     std::smatch sm;
     std::string shapeType("");
     std::string output("");
-    double point1(0.f), point2(0.f);
+    double point1(0.), point2(0.);
     while (tmpString.length() != 0)
     {
-        if (((uint8_t)tmpString.at(0) & 0x80) != 0)
+        if ((static_cast<uint8_t>(tmpString.at(0)) & 0x80) != 0)
         {
             // input is not ascii
             throw std::invalid_argument("input is out of ASCII!");
@@ -163,8 +163,8 @@ std::string Shape::flatten(std::string &shape)
     } STATUS;
     std::string tmpString = shape;
     STATUS status(find_start);
-    double x0(0.f), y0(0.f), x1(0.f), y1(0.f);
-    double x2(0.f), y2(0.f), x3(0.f), y3(0.f);
+    double x0(0.), y0(0.), x1(0.), y1(0.);
+    double x2(0.), y2(0.), x3(0.), y3(0.);
     std::string output("");
     std::string number("");
     std::regex startReg("(-?\\d+\\.?\\d*)\\s+(-?\\d+\\.?\\d*)\\s+b(\\s+)");
@@ -172,7 +172,7 @@ std::string Shape::flatten(std::string &shape)
     std::smatch sm;
     while (tmpString.length() != 0)
     {
-        if (((uint8_t)tmpString.at(0) & 0x80) != 0)
+        if ((static_cast<uint8_t>(tmpString.at(0)) & 0x80) != 0)
         {
             // input is not ascii
             throw std::invalid_argument("input is out of ASCII!");
@@ -271,7 +271,7 @@ std::vector<std::map<std::string, double>> Shape::to_pixels(std::string &shape)
 {
     // Scale values for supersampled rendering
     uint8_t upscale(SUPERSAMPLING);
-    double downscale(0.125f); // 1 / 8
+    double downscale(0.125); // 1 / 8
     std::function<std::pair<double, double>(double, double, std::string &)> flt([&](double x, double y,
                                                                      std::string &input)
     {
@@ -286,8 +286,8 @@ std::vector<std::map<std::string, double>> Shape::to_pixels(std::string &shape)
     double x1(std::get<0>(tmpTuple)), y1(std::get<1>(tmpTuple));
     double x2(std::get<2>(tmpTuple)), y2(std::get<3>(tmpTuple));
 
-    double shift_x(-(x1 - ((int64_t)x1 % upscale)));
-    double shift_y(-(y1 - ((int64_t)y1 % upscale)));
+    double shift_x(-(x1 - (static_cast<int64_t>(x1) % upscale)));
+    double shift_y(-(y1 - (static_cast<int64_t>(y1) % upscale)));
 
     newShape = move(newShape, shift_x, shift_y);
 
@@ -295,7 +295,7 @@ std::vector<std::map<std::string, double>> Shape::to_pixels(std::string &shape)
     double img_width(ceil((x2 + shift_x) * downscale) * upscale);
     double img_height(ceil((y2 + shift_y) * downscale) * upscale);
     std::vector<bool> img_data;
-    img_data.resize(img_width * img_height);
+    img_data.resize(static_cast<size_t>(img_width * img_height));
     fill(img_data.begin(), img_data.end(), false);
 
     // Render shape on image
@@ -303,28 +303,28 @@ std::vector<std::map<std::string, double>> Shape::to_pixels(std::string &shape)
 
     // Extract pixels from image
     std::vector<std::map<std::string, double>> pixels;
-    pixels.reserve((img_height-upscale) * (img_width-upscale));
-    double opacity(0.f);
+    pixels.reserve(static_cast<size_t>((img_height-upscale) * (img_width-upscale)));
+    double opacity(0.);
     for (double y = 0; y <= (img_height - upscale); y += upscale)
     {
         for (double x = 0; x <= (img_width - upscale); x += upscale)
         {
-            opacity = 0.f;
+            opacity = 0.;
             for (uint8_t yy = 0; yy <= (upscale - 1); ++yy)
             {
                 for (uint8_t xx = 0; xx <= (upscale - 1); ++xx)
                 {
-                    if (img_data.at(((y + yy) * img_width) + (x + xx)))
+                    if (img_data.at(static_cast<size_t>(((y + yy) * img_width) + (x + xx))))
                     {
                         opacity += 255;
                     }
                 } // end for xx
             } // end for yy
 
-            if (opacity > 0.f)
+            if (opacity > 0.)
             {
                 std::map<std::string, double> pixel;
-                pixel["alpha"] = (uint8_t)(255 - (opacity * (downscale * downscale)));
+                pixel["alpha"] = static_cast<uint8_t>(255 - (opacity * (downscale * downscale)));
                 pixel["x"] = (x - shift_x) * downscale;
                 pixel["y"] = (y - shift_y) * downscale;
                 pixels.push_back(pixel);
@@ -469,7 +469,7 @@ bool Shape::curve4_is_flat(double x0, double y0,
     {
         if (vecs.at(i).first == 0 && vecs.at(i).second == 0)
         {
-            vecs.erase(vecs.begin() + i);
+            vecs.erase(vecs.begin() + static_cast<int>(i));
             --n;
         }
         else
@@ -631,7 +631,7 @@ void Shape::render_shape(double width, double height,
         if (vy != 0)
         {
             double s((y2 - y) / vy);
-            if (s >= 0.f && s <= 1.f)
+            if (s >= 0. && s <= 1.)
             {
                 ret.push_back(x + s * vx);
                 ret.push_back(y2);
@@ -642,7 +642,7 @@ void Shape::render_shape(double width, double height,
     });
 
     auto tmpTuple(bounding(shape));
-    for (double y = std::max(floor(std::get<1>(tmpTuple)), (double)0.f);
+    for (double y = std::max(floor(std::get<1>(tmpTuple)), static_cast<double>(0.f));
          y <= (std::min(ceil(std::get<3>(tmpTuple)), height) - 1);
          ++y)
     {
@@ -660,7 +660,7 @@ void Shape::render_shape(double width, double height,
             if (cx.size() != 0)
             {
                 row_stops.push_back(std::make_pair(trim(cx.at(0), 0, width),
-                                              std::get<3>(line) > 0.f ? 1.f : -1.f));
+                                    std::get<3>(line) > 0. ? 1. : -1.));
             }
         } //end for i
 
@@ -671,17 +671,17 @@ void Shape::render_shape(double width, double height,
                                                         {
                                                             return (a.first < b.first);
                                                         });
-            double status(0.f), row_index(y * width);
+            double status(0.), row_index(y * width);
             for (size_t i = 0; i < row_stops.size() - 1; ++i)
             {
                 status += row_stops.at(i).second;
-                if (status != 0.f)
+                if (status != 0.)
                 {
                     for (double x = ceil(row_stops.at(i).first - 0.5);
                          x <= (floor(row_stops.at(i + 1).first + 0.5) - 1);
                          ++x)
                     {
-                        image.at(row_index + x) = true;
+                        image.at(static_cast<size_t>(row_index + x)) = true;
                     } // end for x
                 }
             } // end for i
